(fix) db seeds for pgsql

+++ database/seeds/CategoriesTableSeeder.php
@@ -11,8 +11,13 @@ class CategoriesTableSeeder extends Seeder
      */
     public function run()
     {
-        DB::table('products')->truncate();
-        DB::table('categories')->truncate();
+        if (config('database.default') == 'pgsql') {
+            DB::statement('TRUNCATE products CASCADE');
+            DB::statement('TRUNCATE categories CASCADE');
+        } else {
+            DB::table('products')->truncate();
+            DB::table('categories')->truncate();
+        }
         factory(App\Category::class, rand(2, 3))->create()->each(function ($c1) {
             factory(App\Category::class, rand(1, 3))->make()->each(function ($c2) use ($c1) {
                 $c1->appendNode($c2);
